おおまかに完成

+++ app/Http/Controllers/CommentController.php
@@ -29,6 +29,8 @@ class CommentController extends Controller
 
     public function destroy(Comment $comment)
     {
+        $this->authorize('delete', $comment);
+
         $comment->delete();
 
         return back();
